Cambios isb

+++ ExternalServiceMail/Class/AgipEntity.cs
@@ -69,7 +69,7 @@ namespace ExternalService.Class
     {
         public long cuit { get; set; }
         public string nombre { get; set; }
-        public int? isib { get; set; }
+        public long? isib { get; set; }
         public int? cat { get; set; }
         public string calle { get; set; }
         public int? puerta { get; set; }
